refactor: split types/variant_encode.go — Go value encoding into variant_go_encode.go
